feat: fetch all games
